+++ docker/prod/php/bin/init.sh
@@ -15,14 +15,14 @@ done
 
 info "Setting permissions."
 chown www-data:www-data /var/www/openloyalty
-chmod 755 /var/www/openloyalty/var /var/www/openloyalty/var/cache /var/www/openloyalty/var/logs /var/www/openloyalty/app/uploads /var/www/openloyalty/web/uploads /var/www/openloyalty/var/sessions /var/www/openloyalty/var/import
+chmod 755 /var/www/openloyalty/app/var/jwt /var/www/openloyalty/var /var/www/openloyalty/var/cache /var/www/openloyalty/var/logs /var/www/openloyalty/app/uploads /var/www/openloyalty/web/uploads /var/www/openloyalty/var/sessions /var/www/openloyalty/var/import
 
 info "Checking if database is already initialized"
 if [[ `PGPASSWORD="$APP_DB_PASSWORD" psql -h $APP_DB_HOST -U $APP_DB_USER -l -A -t | grep $APP_DB_NAME | wc -l` -gt 0 ]] \
     && ! [[ `PGPASSWORD="$APP_DB_PASSWORD" psql -h $APP_DB_HOST -U $APP_DB_USER -d $APP_DB_NAME -A -t -q -c '\dt' | wc -l` -gt 0 ]]; then
       info "Running initializing scripts..."
-      runuser -s /bin/sh -c 'phing setup' www-data
       chown -R www-data:www-data /var/www/openloyalty
+      runuser -s /bin/sh -c 'phing setup' www-data
       info "Tables are initialized."
 else
     info "Database already initialized."